all working with real weights and random input.
